capa.capabilities: move `has_file_limitations()` from capa.main to the capabilities module
